cleaned up uintptr_t

+++ ex01/include/Serializer.hpp
@@ -1,27 +1,26 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   serializer.hpp                                     :+:      :+:    :+:   */
+/*   Serializer.hpp                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: nige42 <nige42@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/14 11:14:31 by nige42            #+#    #+#             */
-/*   Updated: 2025/03/17 13:17:07 by nige42           ###   ########.fr       */
+/*   Updated: 2025/03/25 10:09:47 by nige42           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #pragma once
 #include <iostream>
 #include <string>
-#include <cstdint>                // For uintptr_t
 #include "./include/Data.hpp"     // structure to hold data
 
 class Serializer {
 
     public:
     
-        static uintptr_t serialize(Data* ptr);
-        static Data* deserialize(uintptr_t raw);
+        static unsigned long serialize(Data* ptr);
+        static Data* deserialize(unsigned long raw);
     
     private:
